requirements.txt
langchain
langchain-google-genai
python-dotenv
gradio
